Added more complex ECMF area test with manually measured expected area
Normalising the existing simple integer test dhists to unit mass and variance results in locations and masses at a range of floating point locations and two ECMFs that are not on the same "grid" at all.
Measurements for expected area between ECMFs made by hand on printed copies of the normalised ECMFs with a grid having x-spacing of 0.02 and y-spacing of 0.01.
Computed area between ECMFs within 1% of manually-measured expected area for both smoothed and unsmoothed cases