Fully vectorised smoothed ECMF area calculation
Vectorised the calculation of bowtie and trapezium segment areas, as well as the code that determines the "shape" of the segment and the intersection point if present. All these functions now operate on a set of vectors of x and y locations.
Brings execution speed for smoothed dhists within 3.2x that of unsmoothed dhists. Given the smoothed code path involves more than 3x the number of vectorised calculations, I think this is "close enough" to parity performance-wise.