Replaced complex bowtie identification / area calculation with simpler approach
Previous bowtie are calculation required identification of point at which the two line intersected, which was computationally expensive.
No using area formula used in existing Python implementation, which doesn't require the intersection point.
Note that this faster formula only holds when y_diff_lower and y_diff_upper have opposite signs and are not both zero.
See issue #21 for verification that this approach is equivalent to the previous approach when the above conditions hold.