Fix package-relative paths to work with R CMD check
Was using system.file(file.path(“inst”, ”extdata", …), package = "netdist") to read
from raw data files in tests. This worked when testing with devotees::test() but
failed when testing with devotees::check(). This appears to be due to the
fact that the files in “inst” are moved to the base directory of the package during
installation. I’m not sure why devtools::test() works both with or without the “inst”
in the relative path from the package root.