added a new method that uses the halfing method
it makes alot fewer calls to the EMD function than the previous method

but we do not seem to be gaining on time

there is two key components to this, one is the overhead of the queue
and the method which can be improved.

but more importantly the code seems to be dominated but the amount of
time it takes to construct all of the offsets. It appears we need to
make this much more efficient.