Add -msse2 flag to ensure correct precision used for all results
Trying to fix Kahan test on Appveyor i386
I don't think this should make a difference as it seems to be about
non-deterministic use of 80-bit precision without the flag, which
should not result in a greater error than using 64-bit precision. If
anything, I'd expect the naiive error to be more likely to fall to 0 in
this case than the Kahan error be non-zero and match the naiive
error (which is what we observe).